Move all testing to testing/test_group
